refactor(awareness): one shared per-terminal sensing leaf, driven by both homes (R9·lib + R9.0)
Replace the maximal `createPulam` assembly with a single shared per-terminal
sensing leaf, `watchTerminalAwareness`, and have BOTH homes drive it: the pulam
daemon by polling its kaval (`awarenessLoop.ts`), kolu-server in-process from its
own spawn/wake/adopt events (`local.ts`). The leaf wires bridgeKavalTaps →
cwd-persist → startAwareness → raw-output activity tap, and bakes in NONE of
discovery / record-seeding / the sink / the served collection / the activity
tracker — all injected by the home.

- createPulam + its two-step start() are dropped: the re-review found that
  boundary complected "sense one terminal" (shared) with each home's discovery /
  seed / sink / collection (per-home). Each home owns its loop; the leaf owns only
  the sensing. `startAwareness`/`bridgeKavalTaps` are no longer exported.
- kolu (R9.0): startAwarenessSensors deletes its inline 4-channel build + its
  startAwareness call and drives the leaf instead, keeping its fold sink + the
  kolu-only exit tap. The leaf takes the already-seeded registry record and NEVER
  re-seeds it — the fold-clobber fix (no clobber on adopt/wake/restore). Sleep/
  wake/adopt/discard + the reader-join are unchanged.
- kolu's served activity goes LIVE (the leaf's byte tap → kolu's tracker, via
  `localTerminalActivity`, replacing `quietActivity`). The client green dot stays
  the local `useTerminalActivity` (it has resize-suppress the server can't
  replicate); the served stream's consumers are remote viewers — a disjoint set,
  not a double-source.
- The structural guard generalizes (`sharedLeaf.assembler.test.ts`): BOTH homes
  rest on the leaf and neither re-rolls the sensing; the differential test
  (daemon path == kolu path == the one leaf) the original R9·lib criterion wanted.
- Docs: remote-terminals-finale.mdx + the r90 SVG reframed around the leaf; a
  History entry records the createPulam→leaf reshape.

pulam-library 73 · pulam 5 (incl. the real-kaval daemon e2e) · kolu-server 149 —
all green; biome lint + typecheck clean; atlas in sync.
